Date picker component work in progress
