+++ components/DatePicker/DatePicker.module.scss
@@ -32,6 +32,11 @@
 
 .calendar {
   margin-right: 1rem;
+  cursor: pointer;
+
+  &:focus {
+    background-color: red;
+  }
 }
 
 .header {
